requirements.txt
Django==4.2.16
Pillow==11.1.0
python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2-binary==2.9.10
whitenoise==6.8.2
gunicorn==23.0.0
